+++ pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
     "datasets>=3.1.0",
     "evaluate>=0.4.3",
     "numpy>=2.0.2",
+    "scikit-learn>=1.5.2",
     "torch>=2.5.1",
 ]
